% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltactrl.R
\name{deltactrl}
\alias{deltactrl}
\title{Calculate the responsiveness of each treatment relative to a control.}
\usage{
deltactrl(data, group_vars, treatment_var, control_label, response_vars)
}
\arguments{
\item{data}{A data frame containing the experimental data.}

\item{group_vars}{A character vector specifying the grouping variables (e.g., c("Genotype", "Block")).}

\item{treatment_var}{Unquoted name of the treatment variable column (e.g., \code{Treatment} or \code{Fertilizer}).}

\item{control_label}{Unquoted label identifying the control treatment level (e.g., \code{Control}).}

\item{response_vars}{A character vector of numeric response variable names (e.g., c("Yield", "Height")).}
}
\value{
A data frame with new columns added for each response variable, prefixed with \code{responsive_}, containing
relative change values compared to the control.
}
\description{
Computes relative change of one or more response variables with respect to a specified control treatment,
within user-defined groups.
}
\details{
The responsiveness is calculated as:
\deqn{(value - mean(control)) / mean(control)}
where "mean(control)" is the group-specific mean of the control level.
NA values in response variables are preserved.#'
}
\examples{
\dontrun{

# upload required package
if(!require(remotes)) install.packages("remotes")
if (!requireNamespace("descriptivestat", quietly = TRUE)) {
remotes::install_github("agronomy4future/deltactrl", force= TRUE)
}
library(remotes)
library(deltactrl)

Fungicide = c("No","No","No","Yes","Yes","Yes")
Fertilizer = c("Control","Fast","Slow","Control","Fast","Slow"),
Yield = c(110.3,119.2,137, 139.5,135.9,156.3
df=data.frame(Fungicide,Fertilizer,Yield)

df1 = deltactrl(
               data = df,
               group_vars = c("Fungicide"),
               treatment_var = Fertilizer,
               control_label = Control,
               response_vars = c("Yield")
)
}

Fungicide  Fertilizer  Yield  responsive_Yield
1 No        Control     110.   NA
2 No        Fast        119.   0.0807
3 No        Slow        137    0.242
4 Yes       Control     140.   NA
5 Yes       Fast        136.   -0.0258
6 Yes       Slow        156.   0.120

* Code source: https://github.com/agronomy4future/deltactrl

}
